Make in_buf and out_buf dynamic
in_buf is fairly straightforward: grow by as much as we want to read, only a few places need to be updated.

out_buf is more tricky because it is grown gradually, as things are being added
to messages and in various places pointers into buffer are held and need to be
updated in case realloc moves the buffer.
To keep track of the pointers we maintain a "buffer context" struct
and take it along with the pointers while building the message.
Most of the places have been modified, with notable exception of
signing and ECDH functions, for these we just take a reasonable guess at
output: 600 bytes, which should be enough for 2048 bit RSA keys or EC
groups. Should be enough for everyone, right? Right?
Well, anyway, we can get back to this later :)

Note: compression is not supported and DTLS mode has not been tested.