Add an option to keep CA verification chain on disk
If MBEDTLS_X509_CA_CHAIN_ON_DISK is defined, the certificate chain can be kept on disk:
it can consist of a single entry with ca_chain_file set, in which case each time verification needs to be performed,
the file is parsed to find only the required certificate (by subject name).
Note that this may need to be done multiple times during verification while traversing the chain provided by the peer.